# Solve Issue #429: Database Connection & Performance Tests