--- a/config/pytest.ini
+++ b/config/pytest.ini
@@ -1,0 +1,15 @@
+[pytest]
+testpaths = tests
+python_files = test_*.py
+python_classes = Test*
+python_functions = test_*
+addopts = -v --tb=short --strict-markers
+pythonpath = src
+markers =
+    integration: integration tests
+    unit: unit tests
+    ml: ML-related tests
+    performance: performance related tests
+filterwarnings =
+    ignore::DeprecationWarning
+    ignore::PendingDeprecationWarning