--- a/data/dbt/models/semantic/metricflow_time_spine.sql
+++ b/data/dbt/models/semantic/metricflow_time_spine.sql
@@ -1,0 +1,6 @@
+-- Time spine model for MetricFlow semantic layer
+{{ config(materialized='table') }}
+
+SELECT 
+    date_day
+FROM {{ ref('time_spine') }}