--- a/data/dbt/neuro_news/macros/incremental_filter.sql
+++ b/data/dbt/neuro_news/macros/incremental_filter.sql
@@ -1,0 +1,12 @@
+{% macro incremental_filter(date_column='published_at_utc', updated_column='updated_at_utc') %}
+  {%- if is_incremental() -%}
+    {{ date_column }}::date > (
+        SELECT COALESCE(MAX({{ date_column }}::date), '1900-01-01'::date) 
+        FROM {{ this }}
+    )
+       OR {{ updated_column }} > (
+        SELECT COALESCE(MAX({{ updated_column }}), '1900-01-01'::timestamp) 
+        FROM {{ this }}
+    )
+  {%- endif -%}
+{% endmacro %}