… APP_PY_COVERAGE_ACHIEVEMENT_REPORT.md
… ISSUE_429_DATABASE_CONNECTION_PERFORMANCE_TESTS_COMPLETE.md
… TEST_COVERAGE_IMPROVEMENT_SUMMARY.md
… artifacts/_mlruns_ci/0/meta.yaml
… artifacts/_mlruns_ci/303720245422693958/meta.yaml
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/artifacts/evaluation_results/tmpw5zy7ttg.csv
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/meta.yaml
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/metrics/avg_answer_f1
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/metrics/avg_mrr
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/metrics/avg_ndcg_at_1
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/metrics/avg_ndcg_at_3
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/metrics/avg_query_time_ms
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/metrics/avg_recall_at_1
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/metrics/avg_recall_at_3
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/metrics/avg_retrieved_docs
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/metrics/queries_evaluated
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/metrics/total_evaluation_time_s
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/params/environment
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/params/experiment_name
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/params/fusion_weight_keyword
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/params/fusion_weight_semantic
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/params/git_branch
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/params/git_sha
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/params/hostname
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/params/k_values
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/params/max_k
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/params/rag_answer_provider
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/params/rag_default_k
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/params/rag_fusion_enabled
… artifacts/_mlruns_ci/390175407796713606/c7f4266d1e6f4fa79a09786bdac419d2/params/reranker_enabled